+++ CHANGES.rst
@@ -15,6 +15,11 @@ Changed
 - UX rework. Submenu addition. :pr:`114`
 - Properly handle LDAP date timezones. :pr:`117`
 
+Fixed
+*****
+
+- CSRF protection on every forms. :pr:`119`
+
 [0.0.22] - 2023-03-13
 =====================